调整重试机制

+++ app/src/main/java/com/drt/moisture/data/source/bluetooth/response/DashboardCorrectDataResponse1.java
@@ -5,7 +5,7 @@ import com.rokyinfo.convert.RkField;
 public class DashboardCorrectDataResponse1 extends CorrectDataResponse{
 
     @RkField(position = 0, length = 1)
-    private byte CmdGroup;
+    private int CmdGroup;
 
     @RkField(position = 1, length = 1)
     private byte Cmd;
@@ -20,7 +20,7 @@ public class DashboardCorrectDataResponse1 extends CorrectDataResponse{
     private long time;
 
     @RkField(position = 9, length = 2)
-    private int temperature1;
+    private short temperature1;
 
     @RkField(position = 11, length = 2)
     private int activity1;
@@ -28,11 +28,11 @@ public class DashboardCorrectDataResponse1 extends CorrectDataResponse{
     @RkField(position = 13, length = 1)
     private int index1;
 
-    public byte getCmdGroup() {
+    public int getCmdGroup() {
         return CmdGroup;
     }
 
-    public void setCmdGroup(byte cmdGroup) {
+    public void setCmdGroup(int cmdGroup) {
         CmdGroup = cmdGroup;
     }
 
@@ -68,11 +68,11 @@ public class DashboardCorrectDataResponse1 extends CorrectDataResponse{
         this.time = time;
     }
 
-    public int getTemperature1() {
+    public short getTemperature1() {
         return temperature1;
     }
 
-    public void setTemperature1(int temperature1) {
+    public void setTemperature1(short temperature1) {
         this.temperature1 = temperature1;
     }
