+++ app/src/main/java/com/drt/moisture/data/source/bluetooth/response/DashboardRecordDataResponse3.java
@@ -5,7 +5,7 @@ import com.rokyinfo.convert.RkField;
 public class DashboardRecordDataResponse3 extends DashboardRecordDataResponse {
 
     @RkField(position = 0, length = 1)
-    private byte CmdGroup;
+    private int CmdGroup;
 
     @RkField(position = 1, length = 1)
     private byte Cmd;
@@ -59,12 +59,12 @@ public class DashboardRecordDataResponse3 extends DashboardRecordDataResponse {
     private short humidity3;
 
     
-    public byte getCmdGroup() {
+    public int getCmdGroup() {
         return CmdGroup;
     }
 
     
-    public void setCmdGroup(byte cmdGroup) {
+    public void setCmdGroup(int cmdGroup) {
         CmdGroup = cmdGroup;
     }
