+++ app/src/main/java/com/drt/moisture/data/source/bluetooth/response/StartMeasureResponse.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 public class StartMeasureResponse {
 
     @RkField(position = 0, length = 1)
-    private byte CmdGroup;
+    private int CmdGroup;
 
     @RkField(position = 1, length = 1)
     private byte Cmd;
@@ -20,11 +20,11 @@ public class StartMeasureResponse {
 
     private int index;
 
-    public byte getCmdGroup() {
+    public int getCmdGroup() {
         return CmdGroup;
     }
 
-    public void setCmdGroup(byte cmdGroup) {
+    public void setCmdGroup(int cmdGroup) {
         CmdGroup = cmdGroup;
     }
