+++ app/src/main/java/com/drt/moisture/data/source/bluetooth/resquest/SetDeviceInfoRequest.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 public class SetDeviceInfoRequest {
 
     @RkField(position = 0, length = 1)
-    private byte CmdGroup;
+    private int CmdGroup;
 
     @RkField(position = 1, length = 1)
     private byte Cmd;
@@ -30,11 +30,11 @@ public class SetDeviceInfoRequest {
     @RkField(position = 53, length = 16)
     private byte[] battery;
 
-    public byte getCmdGroup() {
+    public int getCmdGroup() {
         return CmdGroup;
     }
 
-    public void setCmdGroup(byte cmdGroup) {
+    public void setCmdGroup(int cmdGroup) {
         CmdGroup = cmdGroup;
     }
