+++ app/src/main/java/com/drt/moisture/data/source/bluetooth/resquest/StartMeasureRequest.java
@@ -5,7 +5,7 @@ import com.rokyinfo.convert.RkField;
 public class StartMeasureRequest {
 
     @RkField(position = 0, length = 1)
-    private byte CmdGroup;
+    private int CmdGroup;
 
     @RkField(position = 1, length = 1)
     private byte Cmd;
@@ -31,11 +31,11 @@ public class StartMeasureRequest {
     @RkField(position = 38, length = 1)
     private int index;
 
-    public byte getCmdGroup() {
+    public int getCmdGroup() {
         return CmdGroup;
     }
 
-    public void setCmdGroup(byte cmdGroup) {
+    public void setCmdGroup(int cmdGroup) {
         CmdGroup = cmdGroup;
     }
